Added end to end flow for save employee

+++ src/main/java/com/pkrm/domain/Address.java
@@ -12,7 +12,7 @@ public class Address {
 	private String district;
 	private String state;
 	private String country;
-	private int pinCode;
+	private long pinCode;
 	
 
 }
